requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aioquic==1.2.0
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==5.1.3
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
Flask==3.1.0
frozenlist==1.5.0
graphviz==0.20.3
h11==0.14.0
h2==4.1.0
hpack==4.1.0
hyperframe==6.0.1
idna==3.10
ipaddress==1.0.23
itsdangerous==2.2.0
Jinja2==3.1.6
kaitaistruct==0.10
kamene==0.32
ldap3==2.9.1
mac-vendor-lookup==0.1.12
MarkupSafe==3.0.2
mitmproxy==11.0.2
mitmproxy-windows==0.10.7
mitmproxy_rs==0.10.7
msgpack==1.1.0
multidict==6.4.3
mysqlclient==2.2.7
netifaces==0.11.0
oauthlib==3.2.2
passlib==1.7.4
pillow==11.2.1
propcache==0.3.1
psutil==7.0.0
publicsuffix2==2.20191221
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydivert==2.1.0
PyJWT==2.9.0
pylsqpack==0.3.20
pyOpenSSL==24.3.0
pyparsing==3.2.0
pyperclip==1.9.0
pyreadline3==3.5.4
python-nmap==0.7.1
python-owasp-zap-v2.4==0.1.0
python3-openid==3.2.0
pywin32==310
reportlab==4.4.1
requests==2.32.3
requests-oauthlib==2.0.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
scapy==2.6.1
service-identity==24.2.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sortedcontainers==2.4.0
sqlparse==0.5.2
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2024.2
urllib3==2.2.3
urwid==2.6.16
wcwidth==0.2.13
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.19.0
zaproxy==0.4.0
zstandard==0.23.0
